working sequential operator version. changed triangle_t to be long type, for correct counts on larger graphs

+++ include/Static/TriangleCounting/triangle2.cuh
@@ -10,7 +10,8 @@
 
 namespace hornets_nest {
 
-using triangle_t = unsigned int;
+//using triangle_t = int;
+using triangle_t = unsigned long long int;
 using HornetGraph = gpu::Hornet<EMPTY, EMPTY>;
